Develop: KontextFenster - Content in eigene Klasse verschoben. Debug Controls entfernt. Todo: Klasse Testen.
