+++ w19_memecreator/MainWindow.xaml
@@ -11,14 +11,14 @@
         <Label Content="Templates" HorizontalAlignment="Left" Margin="10,30,0,0" VerticalAlignment="Top"/>
         <StackPanel x:Name="stackP_Timeline" HorizontalAlignment="Left" Height="100" Margin="10,56,0,0" VerticalAlignment="Top" Width="974"/>
 
-        <Border BorderThickness="2" BorderBrush="Black" Height="450" Width="620" Margin="10,161,0,0" VerticalAlignment="Top" HorizontalAlignment="Left">
-            <Canvas x:Name="canvas_Bearbeitungsfenster" HorizontalAlignment="Left" Height="450" Margin="-2" VerticalAlignment="Top" Width="620"/>
+        <Border BorderThickness="2" BorderBrush="Black" Height="450" Width="600" Margin="10,160,0,0" VerticalAlignment="Top" HorizontalAlignment="Left">
+            <Canvas x:Name="canvas_Bearbeitungsfenster" Background="White" HorizontalAlignment="Left" Height="445" Width="595" Margin="0,0,0,0" VerticalAlignment="Top"/>
         </Border>
 
         <!-- Kontextfenster Design -->
         <Border x:Name="border_KontextFenster" BorderThickness="2" BorderBrush="Black" Height="450" Width="344" Margin="636,161,0,0" VerticalAlignment="Top" HorizontalAlignment="Left">
             <ScrollViewer>
-                <Grid x:Name="grid_Kontextfenster" ScrollViewer.CanContentScroll="True" Height="650">
+                <Grid x:Name="grid_Kontextfenster" ScrollViewer.CanContentScroll="True">
 
                 </Grid>
             </ScrollViewer>
@@ -28,7 +28,7 @@
             <MenuItem Header="File">
                 <MenuItem Header="Load"></MenuItem>
                 <MenuItem Header="Save"></MenuItem>
-              <Separator Margin="3,2,-3,2"></Separator>
+                <Separator Margin="3,2,-3,2"></Separator>
                 <MenuItem Header="Exit"></MenuItem>
             </MenuItem>
             <MenuItem Header="Edit">
@@ -37,10 +37,5 @@
                 </MenuItem>
             </MenuItem>
         </Menu>
-
-        <Button x:Name="btn_DEBUG_lbl1" Content="lbl1" HorizontalAlignment="Left" Margin="881,22,0,0" VerticalAlignment="Top" Width="75" Tag="1" Click="btn_DEBUG_lbl1_Click"/>
-        <Label x:Name="lbl_DEBUG_ShowText1" Content="YourTextHere" HorizontalAlignment="Left" Margin="330,202,0,0" VerticalAlignment="Top" Width="157" Height="160" Tag="1"/>
-        <Label x:Name="lbl_DEBUG_ShowText2" Content="YourTextHere" HorizontalAlignment="Left" Margin="330,404,0,0" VerticalAlignment="Top" Width="157" Height="160" Tag="2"/>
-        <Button x:Name="btn_DEBUG_lbl2" Content="lbl2" HorizontalAlignment="Left" Margin="779,22,0,0" VerticalAlignment="Top" Width="75" Tag="2" Click="btn_DEBUG_lbl2_Click"/>
     </Grid>
 </Window>
